Renaming local:sync* to sync*.
