+++ src/Robo/Commands/Sync/FilesCommand.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Acquia\Blt\Robo\Commands\Sync;
+
+use Acquia\Blt\Robo\BltTasks;
+
+/**
+ * Defines commands in the "setup:files*" namespace.
+ */
+class FilesCommand extends BltTasks {
+
+  /**
+   * Synchronize local files from remote (remote --> local).
+   *
+   * @command sync:files
+   */
+  public function syncFiles() {
+    $local_alias = '@' . $this->getConfigValue('drush.aliases.local');
+    $remote_alias = '@' . $this->getConfigValue('drush.aliases.remote');
+    $site_dir = $this->getConfigValue('site');
+
+    $task = $this->taskDrush()
+      ->alias('')
+      ->assume('')
+      ->uri('')
+      ->drush('rsync')
+      ->arg($remote_alias . ':%files')
+      ->arg($this->getConfigValue('docroot') . "/sites/$site_dir/files")
+      ->option('exclude-paths', 'styles:css:js');
+
+    $result = $task->run();
+
+    return $result;
+  }
+
+}
